--- feature/sign/src/main/java/com/qriz/app/feature/sign/model/SignUpEffect.kt
@@ -1,6 +1,0 @@
-package com.qriz.app.feature.sign.model
-
-sealed interface SignUpEffect {
-    data object SignUpComplete : SignUpEffect
-    data object SendAuthEmail: SignUpEffect
-}
