+++ feature/sign/src/main/java/com/qriz/app/feature/sign/signIn/SignInContract.kt
@@ -1,0 +1,28 @@
+package com.qriz.app.feature.sign.signIn
+
+import androidx.compose.runtime.Immutable
+import com.qriz.app.feature.base.UiAction
+import com.qriz.app.feature.base.UiEffect
+import com.qriz.app.feature.base.UiState
+
+@Immutable
+data class SignInUiState(
+    val id: String,
+    val password: String,
+    val showPassword: Boolean,
+    val idErrorMessage: String,
+    val passwordErrorMessage: String,
+) : UiState {
+    companion object {
+        val Default = SignInUiState(
+            id = "",
+            password = "",
+            showPassword = false,
+            idErrorMessage = "",
+            passwordErrorMessage = ""
+        )
+    }
+}
+
+sealed interface SignInUiAction : UiAction
+sealed interface SignInUiEffect : UiEffect
